Issue 38:	NullPointerException in ApiLocation

+++ eveapi/trunk/src/main/java/com/beimin/eveapi/shared/locations/ApiLocation.java
@@ -3,9 +3,9 @@ package com.beimin.eveapi.shared.locations;
 public class ApiLocation {
 	private long itemID;
 	private String itemName;
-	private long x;
-	private long y;
-	private long z;
+	private double x;
+	private double y;
+	private double z;
 
 	public long getItemID() {
 		return itemID;
@@ -23,27 +23,27 @@ public class ApiLocation {
 		this.itemName = itemName;
 	}
 
-	public long getX() {
+	public double getX() {
 		return x;
 	}
 
-	public void setX(long x) {
+	public void setX(double x) {
 		this.x = x;
 	}
 
-	public long getY() {
+	public double getY() {
 		return y;
 	}
 
-	public void setY(long y) {
+	public void setY(double y) {
 		this.y = y;
 	}
 
-	public long getZ() {
+	public double getZ() {
 		return z;
 	}
 
-	public void setZ(long z) {
+	public void setZ(double z) {
 		this.z = z;
 	}
 }
